requirements.txt
streamlit
HEAD
openai
pandas
matplotlib
openpyxl
pyttsx3
psycopg2-binary
python-dotenv
SpeechRecognition
pandas
openai
matplotlib
openpyxl
psycopg2-binary
a2d4b46 (Versión inicial de Cerberu con backend multilenguaje)
